Output tract job loss csvs for Urban Data Portal
Write out tract level job loss CSV in addition to geojson based on user feedback for Urban Data Portal. Note this file was over 100 MB, so not pushed to Github but it is in S3. This CSV along with the county and cbsa sums CSVs will be added to the Urban Data Portal