Fix mistake in BLS data for Aug 7th
The raw BLS all series file was missing several series starting with 01 and 02. So we downloaded in the sub series files for 01 and 02, filtered out any remaining 01 and 02 series in the full series file, and appended the subseries. This seems to be a mistake on BLS' part and we are followig up with them